Don't skip whitespace when decrementing filter string
In incremental search mode, keep an empty string in the list to
represent whitespace when decrementing the filter string.  This avoids
an unnatural skipping of whitespace when pressing delete in
incremental search mode.  In practice, this is all handled by calls
to `deft-filter`.  We do this by passing nil to explicitly pop an item
off of the list, while passing a string, even an empty one, will
replace the car with that string.