Handle filter string properly when toggling search modes
Added an optional RESET argument to deft-filter which, when set,
resets the entire filter string as if the function were called
interactively (e.g., by C-c C-l).  This is in contrast to the
usual non-interactive behavior where deft-filter only operates on
the car.  This patch fixes the filter behavior when toggling
between regex and incremental search.