Add C-o for opening files in another window
Used without a prefix argument (C-o), open the file in another window
without switching there.  When used with a prefix argument (C-u C-o),
switch to the new window.

Also included in this patch: check that newly opened buffers are not
already using `deft-text-mode' before switching to that mode.

Thanks to Juha Ranta <juharanta@me.com> for suggesting this feature.