Try to prevent false title matches in org notes
- With the updated code, if `deft-org-mode-title-prefix` is non-nil,
title match with strictly happen with lines beginning with "^#+TITLE:
". This prevents false extractions of summary in the event that the
"title" string appears elsewhere in the note body too.

Example note:

Some text here that should become part of the summary.
Many people are good people.

Without this patch, only "are good people" shows up in the summary.