Reduce summary line width when there is no fringe
It turns out the issue was not console mode per se, but the absence of a
fringe. In graphical mode, the fringe effectively absorbs the newline
allowing us to use the entire width. In console mode, or when fringe
mode is disabled, we need to add the one-character offset.

This should remove the need for the custom deft-width-offset variable,
so it has been marked as obsolete.